refactor the entire project structure and folder names
