--- src/components/elements/ButtonComponent.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ButtonComponent = (props) => {
-  return <div>ButtonComponent</div>;
-};
-
-export default ButtonComponent;
