+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as NavBarComponent } from "./navbar/NavBarComponent";
+export { default as FooterComponent } from "./footer/FooterComponent";
