--- src/pages/index.js
@@ -1,5 +1,0 @@
-export { default as Homepage } from "./homepage/Homepage";
-export { default as About } from "./aboutpage/About";
-export { default as Contact } from "./contactpage/Contact";
-export { default as Portfolio } from "./portfoliopage/Portfolio";
-export { default as Services } from "./servicespage/Services";
