+++ src/sections/index.js
@@ -1,0 +1,5 @@
+export { default as Hero } from "./hero/Hero";
+export { default as About } from "./about/About";
+export { default as Contact } from "./contact/Contact";
+export { default as Portfolio } from "./portfolio/Portfolio";
+export { default as Services } from "./services/Services";
